nie usuwanie pliku .file

+++ www/class/file_handle.php
@@ -8,11 +8,14 @@ $type =  $_POST['submit'];
 
 //delete files older than 1 month
 $dir = "files/uploaded";
-$timeold = time() - (60 * 60 * 24 * 30);
+$timeold = time() - (60*60*24*30);
 if (file_exists($dir)) {
     $di = new RecursiveDirectoryIterator($dir, FilesystemIterator::SKIP_DOTS);
     $ri = new RecursiveIteratorIterator($di, RecursiveIteratorIterator::CHILD_FIRST);
     foreach ($ri as $file) {
+        if($file->getFilename() == '.file'){
+            continue;
+        }
         if (!$file->isDir() && filemtime($file) < $timeold) {
             unlink($file);
         }
